worked on led matrix code i can send data over serial either character data or pixel data, is going to be fun. i have a python script i am working on with which i can draw on the matrix, future plans include a pixel editor and somehow stream stuff, and also make it store things in it's memory

+++ atmega16-ledmatrix/font.h
@@ -1,7 +1,8 @@
 #ifndef _FONT_H_
 #define _FONT_H_
+
 // Font (ugly imho)
-uint16_t font[760] = {
+uint8_t font[760] = {
   0,0,0,0,0,0,0,0,
   12,12,12,12,0,12,12,0,
   20,20,20,0,0,0,0,0,
